// backend/model/ekyc/tencent/uploadurlresult.ts
// {
//     "UploadUrl": "https://faceid-resource-hk-1258344699.cos.xxx",
//     "ResourceUrl": "https://faceid-resource-hk-1258344699.xxx",
//     "ExpiredTimestamp": 1694079404,
//     "RequestId": "xxx-57e4-xx-a5a1-xx"
// }
export type UploadUrlResult = {
    UploadUrl: string;
    ResourceUrl: string;
    ExpiredTimestamp: number;
    RequestId: string;
}